feat: implement power management components for core s3
The most vital change is in board_m5cores3 where we enable power on the
external Grove ports from USB-C voltage via the AXP2101. Without this,
the power state for DC-DC, ALDO, and BLDO will not be bootstrapped
appropriately to power the ports. SDA will be held low and I2C on the
Port A Grove bus will be unavailable (GPIO1 and 2).

This also disables the display LED backlight. Since we're not using it
in my implementation. I will make this an optional config soon, because
we will want to use it for some Core S3 controllers.

Finally, this implements basic power sensors for the AXP2101, since I
was in there fucking around anyway. We may as well get charging status
and battery percentage.