fix: sync with thermal sensor before getting image data
We have seen numerous issues with the current implementation where the
thermal image will be partially garbage data or completely missing. The
reason is we are trying to get image frame data without coordinating
with the sensor first.

The Melexis driver docs PDF mentions this:

This function waits for a new data to be available for a MLX90640
device with a given slave address. The purpose of the function is to
synchronize with the MLX90640 device so that the data acquisition can be
started right after the data is available. [...]  It is recommended that
the frame synchronization is being used before the very first frame read.